requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
pyyaml>=6.0
requests>=2.28.0
pyzmq>=25.0.0
regex>=2023.0.0
aiohttp>=3.8.0
dataclasses-json>=0.6.0

# Optional dependencies for enhanced functionality
accelerate>=0.20.0
bitsandbytes>=0.41.0
peft>=0.4.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
coverage>=7.0.0
pre-commit>=3.0.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
myst-parser>=2.0.0
